Rename NewRecordListView to original RecordListView
