--- frontend/vre/record/record.list.item.view.mustache
@@ -1,2 +1,0 @@
-<td><input type="checkbox"></td>
-<td><a href="#">{{displayText}}</a></td>
