--- frontend/vre/record/record.list.view.mustache
@@ -1,1 +1,0 @@
-<tbody></tbody>
